Consolidate version-file helpers into VersionGradleFile
Address review on PR #710:
- Move the `VersionGradleFile` object (and its spec) to the `io.spine.gradle`
  package — it reads and parses any `version.gradle.kts`, not only for publishing.
- Fold the file-reading helpers into the object: `VERSION_FILE` -> `NAME`,
  `headVersionFile` -> `contentUnder`, `baseVersionFile` -> `contentInBase`, with
  `gitShow`/`GitResult` now private members.
- Update `CheckVersionIncrement` call sites accordingly.

No behavior change.
